Add dictToTrie
Add function dictToTrie to turn a dict into a trie. The resulting trie
is a dict of dicts of dicts... As a result, no key can be a prefix to
another.

The plan is to have a simple dictionary (word, class) that we
transform into a trie, into which we can then feed a stream of
character that will stop once we reach a full word. The condition
stated above is fulfilled by the lojban's property of being
unambiguous.